# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "polyGraphics 3D Backend",
    "description": "API for 2D-to-3D reconstruction. Default path: **Segment Anything first** (smart background removal), then MapAnything / mesh or Gaussian splatting \u2014 reconstructions consume **masked** views only. `skip_sam_segmentation` is only for pre-cut **RGBA** assets with real transparency.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Direct Uvicorn (e.g. http://127.0.0.1:8000)"
    }
  ],
  "paths": {
...
    "/tripo/image-to-model": {
      "post": {
        "tags": [
          "tripo"
        ],
        "summary": "Post Tripo Image To Model",
        "operationId": "post_tripo_image_to_model_tripo_image_to_model_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_tripo_image_to_model_tripo_image_to_model_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Tripo job accepted; poll GET until completed, then download. Uses Tripo SDK when TRIPO_API_KEY is set; otherwise returns a placeholder GLB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripoImageToModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tripo/image-to-model/{tripo_job_id}": {
      "get": {
        "tags": [
          "tripo"
        ],
        "summary": "Get Tripo Image To Model",
        "operationId": "get_tripo_image_to_model_tripo_image_to_model__tripo_job_id__get",
        "parameters": [
          {
            "name": "tripo_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tripo Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripoImageToModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tripo/image-to-model/{tripo_job_id}/download": {
      "get": {
        "tags": [
          "tripo"
        ],
        "summary": "Download Tripo Model",
        "operationId": "download_tripo_model_tripo_image_to_model__tripo_job_id__download_get",
        "parameters": [
          {
            "name": "tripo_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tripo Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_post_tripo_image_to_model_tripo_image_to_model_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image",
            "description": "Single product/object photo (JPEG, PNG, or WebP)."
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_post_tripo_image_to_model_tripo_image_to_model_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TripoImageToModelResponse": {
        "properties": {
          "tripo_job_id": {
            "type": "string",
            "title": "Tripo Job Id",
            "description": "PolyGraphics Tripo job id (poll this endpoint)."
          },
          "status": {
            "$ref": "#/components/schemas/TripoJobStatus"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url",
            "description": "Relative API path to download the GLB when status is completed."
          },
          "tripo_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tripo Task Id",
            "description": "Tripo cloud task id when known."
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "tripo_job_id",
          "status"
        ],
        "title": "TripoImageToModelResponse"
      },
      "TripoJobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "processing",
          "completed",
          "failed"
        ],
        "title": "TripoJobStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Google sign-in access token from POST /auth/google or GET /auth/google/callback."
      }
    }
  }
}